+++ src/main.rs
@@ -123,11 +123,11 @@ fn main() {
 
     if ctx.gpu_enabled {
         println!("Running GPU-based training...\n");
-        let _ = run_neural_net::<GpuTensor<f64>>();
+        let _ = run_neural_net::<GpuTensor<f32>>();
         println!("\n✓ All training tasks completed");
     } else {
         println!("Running CPU-based training...\n");
-        let _ = run_neural_net::<CpuTensor<f64>>();
+        let _ = run_neural_net::<CpuTensor<f32>>();
         println!("\n✓ All training tasks completed");
     }
 }
